regby-rostov/requirements.txt
django==2.1.1
djangorestframework==3.8.2
djangorestframework-gis==0.13
requests==2.19.1
lxml==4.2.5
psycopg2==2.7.5
psycopg2-binary==2.7.5
postgis==1.0.4
Pillow==5.2.0
django-filter==2.0.0
numpy